pyproject.toml
[project]
name = "posthog-ducklake-loader"
version = "0.1.0"
description = "Load PostHog event data from S3 into DuckLake using Apache Spark"
requires-python = ">=3.10"
dependencies = [
    "pyspark>=4.0.0",
    "duckdb>=1.3.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "boto3>=1.34.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
